Title: 세 실수의 반올림, Time: 141ms, Memory: 11MB, Status: Passed - CodeTree

+++ 240926/세 실수의 반올림/rounding-of-three-actual-numbers.java
@@ -2,9 +2,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        float a = sc.nextDouble();
-        float b = sc.nextDouble();
-        float c = sc.nextDouble();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
+        double c = sc.nextDouble();
         System.out.printf("%.3f\n%.3f\n%.3f",a,b,c);
     }
 }
